requirements.txt
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
pillow==11.0.0
streamlit==1.41.1
